fix: Name spaced additional properties type
