+++ app/graphql/types/stop_point_type.rb
@@ -1,7 +1,7 @@
 module Types
   class StopPointType < Types::BaseObject
     field :accessibilitySummary, String, null: true
-    field :additionalProperties, [Types::AdditionalPropertiesType, null: true], null: true
+    field :additionalProperties, [Types::Tfl::Entities::AdditionalPropertiesType, null: true], null: true
     field :children, [Types::Tfl::Entities::PlaceType, null: true], null: true
     field :childrenUrls, [String, null: true], null: true
     field :commonName, String, null: true
